penyesuaian html di view/admin

+++ application/views/admin.php
@@ -31,7 +31,7 @@
             <td><?= $student['tanggal_lahir'] ?></td>
             <td>
                 <a href="<?= base_url();?>Admin/edit/<?= $student['id']; ?>">Edit</a> | 
-                <a href="<?= base_url();?>Admin/delete/<?= $student['id']; ?>">Delete</a>
+                <a onclick="return confirm('Bener Mau Hapus nii?')" href="<?= base_url();?>Admin/delete/<?= $student['id']; ?>">Delete</a>
             </td>
         </tr>
         <?php endforeach; ?>
